requirements.txt
facebook-business>=22.0.0
google-ads>=25.0.0
google-genai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
jinja2>=3.1.0
